Remove the html5 pushState simulation default.
Regardless of the merits of whether it be on or off,
enabling it in this framework plugin is confusing b/c
there's nothing inherent to Angular per se that should
suggest users will assume it's turned on.

If it weren't for the trade-offs with pushState simulation
and API Proxy working in tandem, this would be on by
default in Lineman core, but alas.

Fixes #4